backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
redis==4.6.0
requests==2.31.0
beautifulsoup4==4.12.2
textblob==0.17.1
python-dotenv==1.0.0
schedule==1.2.0
gunicorn==21.2.0
feedparser==6.0.10
lxml==6.0.0
